+++ wolfssl/internal.h
@@ -6287,14 +6287,12 @@ WOLFSSL_LOCAL int cipherExtraData(WOLFSSL* ssl);
 
     WOLFSSL_LOCAL void WriteSEQ(WOLFSSL* ssl, int verifyOrder, byte* out);
 
-#if defined(WOLFSSL_TLS13) && (defined(HAVE_SESSION_TICKET) || !defined(NO_PSK))
 #ifdef WOLFSSL_32BIT_MILLI_TIME
-    WOLFSSL_LOCAL word32 TimeNowInMilliseconds(void);
+    WOLFSSL_API word32 TimeNowInMilliseconds(void);
 #else
-    WOLFSSL_LOCAL sword64 TimeNowInMilliseconds(void);
+    WOLFSSL_API sword64 TimeNowInMilliseconds(void);
 #endif
 
-#endif
 WOLFSSL_LOCAL word32  LowResTimer(void);
 
 WOLFSSL_LOCAL int FindSuiteSSL(const WOLFSSL* ssl, byte* suite);
